\name{create_trees}
\alias{create_trees}
\title{Create phylogenetic tree for test purpose}
\usage{
create_trees(n_trees, n_species, q_size, unique=FALSE)
}
\arguments{
    \item{n_trees}{The number of trees to create.}
    \item{n_species}{The number of tips.}
    \item{q_size}{The number of states of the character.}
    \item{unique}{Should the trees be unique? Default to FALSE.}
}
\description{
    Description
}
\details{
    Details
}
\value{
    Returns
}

\references{
    References
}

\author{Diego Pasqualin \email{dpasqualin@inf.ufpr.br}}
\seealso{
    See Also
}
\keyword{phylogenetics}
%\keyword{simulation}
