\name{sfreemap.diff}
\alias{sfreemap.diff}
\title{Simulation free stochastic character maps on a phylogenetic tree}
\usage{
sfreemap.diff(a, b)
}
\arguments{
    \item{a}{A result of a simulation.map execution with Q='mcmc'.}
    \item{b}{A result of a simulation.map execution with Q='mcmc'.}
}
\description{
    Description
}
\details{
    Details
}
\value{
    Returns
}

\references{
    References
}

\author{Diego Pasqualin \email{dpasqualin@inf.ufpr.br}}
\seealso{
    See Also
}
\keyword{phylogenetics}
%\keyword{simulation}