\name{sfreemap.test.perf}
\alias{sfreemap.test.perf}
\title{Simulation free stochastic character maps on a phylogenetic tree}
\usage{
sfreemap.test.perf(tree_seq, species_seq, q_size_seq
								, n_tests=5, parallel=TRUE
								, prog='sfreemap'
								, message=TRUE
                                , file=NULL)
}
\arguments{
    \item{tree_seq}{trololo}
    \item{species_seq}{trololo}
    \item{q_size_seq}{trololo}
    \item{n_tests}{trololo}
    \item{parallel}{trololo}
    \item{prog}{trololo}
	\item{message}{trololo}
	\item{file}{trololo}

}
\description{
    Description
}
\details{
    Details
}
\value{
    Returns
}

\references{
    References
}

\author{Diego Pasqualin \email{dpasqualin@inf.ufpr.br}}
\seealso{
    See Also
}
\keyword{phylogenetics}
