\name{simmap.mean}
\alias{simmap.mean}
\title{Simulation free stochastic character maps on a phylogenetic tree}
\usage{
simmap.mean(mtrees)
}
\arguments{
    \item{mtrees}{A result of a simmap.map execution with Q='mcmc'.}

}
\description{
    Description
}
\details{
    Details
}
\value{
    Returns
}

\references{
    References
}

\author{Diego Pasqualin \email{dpasqualin@inf.ufpr.br}}
\seealso{
    See Also
}
\keyword{phylogenetics}
%\keyword{simulation}
