\name{simulation.data}
\alias{simulation.data}
\title{Simulation free stochastic character maps on a phylogenetic tree}
\usage{
simulation.data(tree)
}
\arguments{
    \item{tree}{A result of a simulation.map execution with Q='mcmc'.}

}
\description{
    Description
}
\details{
    Details
}
\value{
    Returns
}

\references{
    References
}

\author{Diego Pasqualin \email{dpasqualin@inf.ufpr.br}}
\seealso{
    See Also
}
\keyword{phylogenetics}
%\keyword{simulation}